requirements.txt
allure-behave
behave
allure-python-commons
allure-combine
Appium-Python-Client
browserstack-local
Python-BrowserStack
paver
async-generator
attrs
certifi
cffi
chardet
charset-normalizer
cryptography
decorator
defusedxml
docutils
h11
idna
jira
location
oauthlib
outcome
parse
parse-type
Paver
pbr
pluggy
pycparser
PyJWT
PySocks
PyYAML
requests
requests-oauthlib
requests-toolbelt
selenium
six
sniffio
sortedcontainers
trio
trio-websocket
urllib3
wsproto
xlrd
zipp